requirements.txt
# Arya Voice Agent - Production Dependencies (PHONE-MODE)
# Optimized for Twilio telephony - ~200MB disk footprint

# Core Framework
fastapi>=0.115.0
uvicorn>=0.30.0

# AI/ML (Cloud-based - no heavy local models)
langchain>=0.3.0
langchain-groq>=0.2.0
langgraph>=0.2.0

# Vector Search & Embeddings (Cloud-Powered)
numpy>=1.26.0

# Text & Audio Processing
inflect>=7.2.1
edge-tts>=6.1.12
av>=13.0.0
webrtcvad-wheels>=2.0.11

# Security & Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1
passlib>=1.7.4

# Persistent Memory Storage
redis[hiredis]>=5.0.8

# Communication & SDKs
twilio>=9.3.0
groq>=0.11.0
python-dotenv>=1.0.1

# Optional: For WebRTC web mode (adds ~500MB)
# pip install fastrtc gradio librosa numba
